# api.yaml
openapi: 3.1.0
info:
  title: 库存管理系统接口
  version: 0.0.1
servers:
  - url: 'https://development.gigantic-server.com/v1'
    description: Development server
  - url: 'https://staging.gigantic-server.com/v1'
    description: Staging server
  - url: 'https://api.gigantic-server.com/v1'
    description: Production server
security:
  - oauth2: []
  - bearerAuth: []
paths:
...
  /procurement/order:
    post:
      summary: 创建注文书
      operationId: createOrder
      tags:
        - 采购
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderReq'
      responses:
        '200':
          description: 成功创建注文书
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: 请求参数错误
        '404':
          description: 客户不存在
        '409':
          description: 注文明细不合法
        '500':
          description: 服务器内部错误
  '/procurement/order/{id}':
    get:
      summary: 注文详情
      operationId: getOrder
      tags:
        - 采购
      parameters:
        - name: id
          in: path
          required: true
          description: 注文ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功获取注文详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: 注文不存在
    put:
      summary: 更新注文书
      operationId: updateOrder
      tags:
        - 采购
      parameters:
        - name: id
          in: path
          required: true
          description: 注文ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderReq'
      responses:
        '200':
          description: 成功更新注文书
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: 请求参数错误
        '404':
          description: 注文不存在
        '409':
          description: 注文状态不允许修改
    delete:
      summary: 取消注文书
      operationId: cancelOrder
      tags:
        - 采购
      parameters:
        - name: id
          in: path
          required: true
          description: 注文ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功取消注文书
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: 注文不存在
        '409':
          description: 当前状态不允许取消
  '/procurement/order/{id}/submit':
    post:
      summary: 提交注文书
      operationId: submitOrder
      tags:
        - 采购
      parameters:
        - name: id
          in: path
          required: true
          description: 注文ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功提交注文书
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: 注文不存在
        '409':
          description: 当前状态不允许提交
        '500':
          description: 服务器内部错误
  '/procurement/order/{id}/send':
    post:
      summary: 下发注文书
      operationId: sendOrder
      tags:
        - 采购
      parameters:
        - name: id
          in: path
          required: true
          description: 注文ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功下发注文书
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: 注文不存在
        '409':
          description: 当前状态不允许下发
        '500':
          description: 服务器内部错误
  '/procurement/order/{id}/complete':
    post:
      summary: 完成注文书
      operationId: completeOrder
      tags:
        - 采购
      parameters:
        - name: id
          in: path
          required: true
          description: 注文ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功完成注文书
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: 注文不存在
        '409':
          description: 当前状态不允许完成
        '500':
          description: 服务器内部错误
  '/procurement/order/{id}/cancel':
    post:
      summary: 取消注文书
      operationId: cancelOrderAction
      tags:
        - 采购
      parameters:
        - name: id
          in: path
          required: true
          description: 注文ID
          schema:
            type: integer
      responses:
        '200':
          description: 成功取消注文书
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: 注文不存在
        '409':
          description: 当前状态不允许取消
        '500':
          description: 服务器内部错误
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        password:
          tokenUrl: /auth/login
          scopes: {}
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Address:
      type: object
      description: 地址
      properties:
        postalCode:
          type: string
          description: 邮政编码
        detailAddress1:
          type: string
          description: 详细地址1
        detailAddress2:
          type: string
          description: 详细地址2
...
    Order:
      type: object
      properties:
        id:
          type: integer
          description: 注文ID
        orderNumber:
          type: string
          description: 注文编号
        status:
          type: string
          description: 状态
          enum:
            - draft
            - requested
            - sent
            - completed
            - cancelled
        customer:
          type: object
          description: 客户信息
          properties:
            id:
              type: integer
              description: 客户ID
            name:
              type: string
              description: 客户名称
          additionalProperties: false
        deliveryDueDate:
          type: string
          format: date
          description: 纳期
        deliveryAddress:
          $ref: '#/components/schemas/Address'
        contactName:
          type: string
          description: 负责人姓名
        contactPhone:
          type: string
          description: 联系电话
        items:
          type: array
          description: 注文明细列表
          items:
            $ref: '#/components/schemas/OrderItem'
        totalAmount:
          type: number
          format: float
          description: 总金额（未含税）
        currency:
          type: string
          description: 货币，仅支持JPY
          enum:
            - JPY
          default: JPY
        notes:
          type: string
          description: 备注
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
      additionalProperties: false
    OrderItem:
      type: object
      properties:
        id:
          type: integer
          description: 注文明细ID
        product:
          type: object
          description: 商品信息
          properties:
            id:
              type: integer
              description: 商品ID
            name:
              type: string
              description: 商品名称
            sku:
              type: string
              description: 商品SKU
          additionalProperties: false
        quantity:
          type: number
          format: float
          description: 订购数量
        unit:
          type: string
          description: 计量单位
        unitPrice:
          type: number
          format: float
          description: 单价（未含税）
        currency:
          type: string
          description: 货币，仅支持JPY
          enum:
            - JPY
          default: JPY
        lineAmount:
          type: number
          format: float
          description: 明细金额（未含税）
        note:
          type: string
          description: 明细备注
      additionalProperties: false
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
          description: 商品ID
        quantity:
          type: number
          format: float
          description: 订购数量
        unit:
          type: string
          description: 计量单位
        unitPrice:
          type: number
          format: float
          description: 单价（未含税）
        currency:
          type: string
          description: 货币，仅支持JPY
          enum:
            - JPY
          default: JPY
        note:
          type: string
          description: 明细备注
      additionalProperties: false
      required:
        - productId
        - quantity
        - unit
        - unitPrice
    CreateOrderReq:
      type: object
      properties:
        customerId:
          type: integer
          description: 客户ID
        deliveryDueDate:
          type: string
          format: date
          description: 纳期
        deliveryAddress:
          $ref: '#/components/schemas/Address'
        contactName:
          type: string
          description: 负责人姓名
        contactPhone:
          type: string
          description: 联系电话
        items:
          type: array
          description: 注文明细
          items:
            $ref: '#/components/schemas/OrderItemInput'
          minItems: 1
        notes:
          type: string
          description: 备注
      additionalProperties: false
      required:
        - customerId
        - deliveryDueDate
        - deliveryAddress
        - contactName
        - contactPhone
        - items
    UpdateOrderReq:
      type: object
      description: 注文更新请求，可部分更新字段
      properties:
        deliveryDueDate:
          type: string
          format: date
          description: 纳期
        deliveryAddress:
          $ref: '#/components/schemas/Address'
        contactName:
          type: string
          description: 负责人姓名
        contactPhone:
          type: string
          description: 联系电话
        items:
          type: array
          description: 注文明细
          items:
            $ref: '#/components/schemas/OrderItemInput'
          minItems: 1
        notes:
          type: string
          description: 备注
      additionalProperties: false
      minProperties: 1
...